Separate out authorship.rs
